arm64: Fix call without frame pointer save/setup
There are two scenarios of saving the stack pointer (sp) to the base
pointer (bp) in arm64 code:
1. When pre-index is not used, arm64 employs the 'add' instruction to
   save sp to bp.
2. If pre-index is used, then the 'mov' instruction is used to save
   sp to bp.

An example of each scenario:

1. Without pre-index:
   stp x29, x30, [sp, #64]
   add x29, sp, #0x40

2. With pre-index:
   stp x29, x30, [sp, #-16]!
   mov x29, sp

Objtool handles the 'mov' instruction correctly. However, for the 'add'
instruction, objtool doesn't save sp to bp and it doesn't handle offsets
appropriately, resulting in a warning.

Fixes #1