arm64: Resolves most unreachable instruction warnings.
Most of the warnings come from new_insn not being found.
The handle_group_alt function sets new_insn not found to nop,
but because arch_handle_alternative sets skip_alt to true for new_insn not found,
the code fails to execute into the handle_group_alt function.

But there are still other unreachable instruction warnings that need to be debugged.

Fixes #2