objtool: bug: Add reachable annotation to warning macros
openEuler inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I8B8HU
CVE: NA

-------------------------------

WARN* and BUG* both use brk #0x800 opcodes and the distinction is
provided by the contents of the bug table. This table is not accessible
to objtool, so add an annotation to WARN* macros to let objtool know
that brk handler will return an resume the execution of the instructions
following the WARN's brk.

Signed-off-by: Julien Thierry <jthierry@redhat.com>
Signed-off-by: Chen Zhongjin <chenzhongjin@huawei.com>
Signed-off-by: Yihuan Pan <xun794@gmail.com>